+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- (InputField): Fix props not being passed to inner input element
+
 # v2.0.0-dev.2
 
 - (ITheme): Update breakpoints for media queries; export numeric values from `layout/responsive/breakpoints.ts`